requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6

gradio==4.44.0

numpy==1.23.5
pandas==2.0.3
scipy==1.10.1
scikit-learn==1.3.2
matplotlib==3.8.4
seaborn==0.13.2

tensorflow==2.12.0
keras==2.12.0
h5py
protobuf==3.20.3

torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1

opencv-python==4.7.0.72
pillow

transformers==4.43.3
sentence-transformers==2.7.0
huggingface-hub
tokenizers
safetensors

langchain==0.3.7
langchain-community
langchain-openai
langchain-huggingface
langgraph

openai
tiktoken

faiss-cpu

SQLAlchemy
passlib
python-multipart

python-dotenv
requests
httpx

sendgrid
elevenlabs
pydub

PyMuPDF
reportlab

tqdm
rich
psutil